Move History related type alias imports
