+++ client/src/components/History/Multiple/MultipleViewList.vue
@@ -3,11 +3,11 @@ import { computed, type Ref, ref } from "vue";
 //@ts-ignore missing typedefs
 import VirtualList from "vue-virtual-scroll-list";
 
+import type { HistorySummary } from "@/api";
 import { copyDataset } from "@/api/datasets";
 import { useAnimationFrameResizeObserver } from "@/composables/sensors/animationFrameResizeObserver";
 import { useAnimationFrameScroll } from "@/composables/sensors/animationFrameScroll";
 import { Toast } from "@/composables/toast";
-import type { HistorySummary } from "@/stores/historyStore";
 import { useHistoryStore } from "@/stores/historyStore";
 import localize from "@/utils/localization";
